Resolved compiler errors from splitting CAN and UART to their own files
Because:
- This increases readability of main.c

This is a part of the effort to refactor the MCU code to be more
readable. All CAN and UART related functions and variables were put into
their own header and source files. Additionally, globals.h and global.c
were made to contain functions and variables that will be used across
multiple source files.